refactor: move CompilerTop into the compilertop package
Final step of the compiler-top breakup. The class itself stays
as-is — it's already small (~320 lines of lifecycle + frame
snapshotting). Just relocates the file via git mv (history
preserved) so the `util/` package no longer references compiler-top
internals; every component of the TUI now lives under
`tools/compilertop/`.

  - Package: `ca.uwaterloo.flix.util` → `ca.uwaterloo.flix.tools.compilertop`.
  - Dropped four redundant same-package imports (Layout, Profiler,
    FrameState, Renderer) — these are top-level types in the
    compilertop package and accessible without explicit import from
    a sibling file. Wildcard member imports (Aggregation.*, Ansi.*,
    Formatting.*, Model.*) stay since they bring object members
    into scope.
  - Flix.scala: extended the existing `Profiler` import to also
    import `CompilerTop`, since the `util.*` wildcard no longer
    covers it.

Pure code move, no behavior change.
